INTR-304: Refactor code
