+++ src/utils/searchIndexUtils.ts
